ad9694_500ebz: Change reference clock location
The FPGA reference clock input on the converter board is single ended, but
the clock outputs from the clock board are differential. So we can't
directly connect them.

Instead of taking the reference clock from the converter board directly
connect it to the FPGA board where a differential reference clock input is
available on the REFCLK_SMA SMA connectors.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>